refactor(collections): extract CreateOrUpdateModal to separate component
- Extract CreateOrUpdateModal from collections.list.tsx to new file create-or-update-collection-modal.tsx
- Add modal state management with atom in collection.ts
- Update layout-modals.tsx to include the new modal
- Replace local state management with global atom pattern for better reusability
